Fix catholic form page when validations fail with attachment.
When a jobseeker attaches a baptism certificate and submits the form
without filling onther required fields, the baptism certificate exists
as a temp file, but is not attached.

Once the page renders to display the validation errors, it triggers an
exception when tried to call attachment methods over the certificate to
generate a link to it.

Resolving it by only calling these methods when the file is attached.