Fix: Remove flexible working pattern from profiles
The flexible working pattern present in some jobseeker profile
preferences is a non-valid value that breaks the prefilling of Job
Application data from the profile.

When the "flexible" value attempts to write in the Job Application, is
not a valid value for the Job Application enumerable and raises an
exception.

As a consequence, it breaks the flow for these users when trying to
apply for a vacancy.