Fix wrong timestamp displayed for jobseekers
When the jobseekers see their list of Job Applications in their account,
the wrong timestamp is being displayed near the "Submitted"
applications.

It should be displaying the submitted_at, not the updated_at.

This caused confusion for jobseekers that reported it thinking there is
a bug that made them miss the vacancy application timeline even when it
was submitted in time.