Fix: Remove flexible working pattern from profiles (#7674)
## Trello card URL
- Fixes this Sentry error:
https://teaching-vacancies.sentry.io/issues/6295531964

## Changes in this PR:

The flexible working pattern present in some jobseeker profile
preferences is a non-valid value that breaks the prefilling of Job
Application data from the profile.

When the "flexible" value attempts to write in the Job Application, is
not a valid value for the Job Application enumerable and raises an
exception.

As a consequence, it breaks the flow for these users when trying to
apply for a vacancy.

## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

**Note: This fixes an unintended impact from the enumerable options
change.**

Does this change affect any of these integrations?
~- [ ] DfE Analytics platform~
~- [ ] Legacy imports mappings~
~- [ ] DWP Find a Job export mappings~
~- [ ] Publisher ATS API (may require mapping updates or API
versioning)~

### User Experience & Data Integrity

Could this change impact:
~- [ ] Existing subscription alerts (will legacy subscription search
filters break?)~
~- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)~
~- [ ] In-progress drafts for Vacancies or Job Applications~
</details>