Fix catholic form page when validations fail with attachment. (#7667)
## Trello card URL

- https://teaching-vacancies.sentry.io/issues/6437668574

## Changes in this PR:

When a jobseeker attaches a baptism certificate and submits the form
without filling other required fields, the baptism certificate exists as
a temp file but is not attached.

Once the page renders to display the validation errors, it triggers an
exception when trying to call attachment methods over the certificate to
generate a link to it.

Resolving it by only calling these methods when the file is attached.