Fix wrong timestamp displayed for jobseekers (#7675)
## Changes in this PR:

When the jobseekers see their list of Job Applications in their account,
the wrong timestamp is being displayed near the "Submitted"
applications.

It should be displaying the submitted_at, not the updated_at.

This caused confusion for jobseekers that reported it thinking there is
a bug that made them miss the vacancy application timeline even when it
was submitted in time.

## Screenshots of UI changes:

### Fixes the wrong source of data being used in the highlighted line:

![image](https://github.com/user-attachments/assets/0a201ca4-f1be-4ccc-bd3c-795a8385a06c)


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>