Change TRN question (#7676)
This change removes the has_teacher_reference_number question from the
jobseeker profile and job application forms and allows users instead
just to enter the TRN into a textbox.

Due to this change the validations have changed. Previously if the user
answered "yes" to `has_teacher_reference_number` or
`qualified_teacher_status` then we would force the jobseeker to enter
their TRN. Now that `has_teacher_reference_number` is no longer a part
of the form, the jobseeker is only forced to enter their TRN if they
answer yes to `qualified_teacher_status`.