Google index vacancies (#7689)
This change ensures that we when update and delete vacancies from our
ATS API we also update the indexing of these vacancies with google.

While working on this I noticed that we do not index vacancies when they
are published. We index them when they're create/updated if this is done
on the day of publishing, however if they are created/updated prior to
the publishing date we do not index them when they go live. To solve
this problem I have added a scheduled job which indexs vacancies which
are published each day.