Fix error when no "vacancy:" attribute is provided
This was one quite hard to discover, because the request spec
automatically add a "vacancy:" parameter.

But the behaviour is different when you just send an empty body in a
real API request. And that was bringing a 500 error with no context.

Tweaked the tests to be able to reproduce a real "no 'vacancy' param
comes in the request" that would come from the API request.

Once reproduced that, adapted the code to return a schema validation
error based on the lack of 'vacancy' attribute.