Migrate now defunct job preferences role "other_teaching_support" to "other_support" (#7718)
This PR adds a rake task to migrate the "other_teaching_support" role in
existing jobseeker job preferences to "other_support" as we no longer
support the "other_teaching_support" role.

Note: we removed the ability for jobseeker to add this role to their job
preferences a while ago but missed migrating the 792 existing job
preferences that contain this role.