Tighten vacancy duplication validation
Make vacancy duplication validation less prone to false positives.

Some of our ATS clients are trying to publish different positions for
the same schools but with slightly differences (like different working
patterns or salaries).

They should not be detected as a duplicated vacancy in our system.

Adding extra fields to check for during the duplication validation.