Tighten vacancy duplication validation (#7768)
## Trello card URL
-
https://trello.com/c/S6di5g4q/1831-resolve-issues-raised-by-eteach-on-the-publishers-ats-api

## Changes in this PR:

Make vacancy duplication validation less prone to false positives.

Some of our ATS clients are trying to publish different positions for
the same schools but with slight differences (like different working
patterns or salaries).

They should not be detected as a duplicate vacancy in our system.

Adding extra fields to check for during the duplication validation.