Exclude /ats-api from http basic authentication (#7777)
So the API requests do not require to have HTTP user/password. 
Any client would require a correct API KEY to be able to use it.

Raised since needing to bypass it for Staging made me realise there is
no host/environment where the HTTP AUTH for the API endpoint would be
needed.