Use Geocoding Coordinates for "within UK validator"
Doing separate calls to Geocoder.search to check the result country for
validating location inputs, while being the "proper" way to check a
location term country, has some issues in our usage context:

1. Some users reported the validation failing when searching some term
   suggested by the location autocomplete. The location should be valid.
2. Each search hits the Google Geocode API, increasing the service
   costs.

The changes propose a simplification: Reusing the location coordinates
queried and cached by our service for job searches.

Since our Google API coordinate queries are restricted to "within the
uk", we can use the lack of coordinates or the lack or the "default"
response for a missing location returned by the API as an indication to
classify a location as within/outside of the UK.

This bring some advantages:
- Aligning results between the location search and the validation.
  To avoid inconsistencies as the search case reported by some users.
- Rely on cached values. Reducing our API usage costs.

But also some a downside:
- The validation gets more "relaxed".
  Since relying on coordinates is more prone to allow non UK locations
  to return som UK coordinate value and pass the validation.

Overall I think that is a good tradeoff to have. Since this validation
is not critical and we rather have some people passing it and not
getting results/alerts than some legit location searches failing on the
form validation, plus the cost reduction.