Bugfix: Don't update publish_on when invalid/blank
There was a bug on our system allowing published vacancies to have no
publish_on date after being already published.

The bug was triggered when a publisher visited the important dates form
to change the expiration date for an already published vacancy.

The publish_on date for an already published vacancy in the past/today
is not editable in the form, but the "publish_on" was being set
nevertheless in the form to "nil" and saved into DB.

So each time a published vacancy (not scheduled for future publishing)
got their expiry date/time updated, the "publish_on" was being wiped.

Since we are introducing a model validation to not allow published
vacancies without a "publish_on" date present, a system test for the
expiry date/time update flow over a published vacancy started failing.
This helped to identify that the cause of the test failure was the bug,
attempting to save publish_on as nil upon expiry date/time update.

With the bugfix, the form will return "publish_on" as parameter to save
ONLY when publish_on is editable, present and a Date.