Fix bug that removed the publish date on already published vacancies (#7802)
## Trello card URL
-
https://trello.com/c/LQnXrxjT/1878-bug-some-non-discarded-published-vacancies-have-no-publishon-timestamp

## Changes in this PR:

### Fix bug removing `publish_on` from already published vacancies.
The implementation of the important dates form was causing the
publish_on date to be removed when publishers updated the vacancy
expiration date and time, even though the `publish_on` field is not
editable at that stage.

The parameter was still being set (to `nil`) and updated into the
Vacancy in DB upon form save.

We're adding a rake task to backfill any vacancy historically affected
by this bug (has been like this for multiple years, since it was built
:/ ). Fortunately, only 70 or so vacancies seemed to be affected by this
(many of those already expired).

### Use a model callback instead of a method override

When converting a published vacancy back into a draft, instead of
overriding the status change method, we will use a more controlled
callback method to reset the publish_on value only after the status
change and before saving the updated value.

### Add validation to ensure the publish date is present

Only enforced on published vacancies. A Vacancy shouldn't be able to be
published unless they have a publish_on date.

## Screenshot of form page with bug:
![Screenshot From 2025-06-10
15-33-38](https://github.com/user-attachments/assets/049bc709-d3ad-41dc-9f3f-d24fd42c3f8e)


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [x] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [x] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>