Bugfix: Correctly query whether vacancy is for a religious school (#7901)
We were getting complaints about pdf downloads not working. It looks
like the issue was that we were trying to render the religious
information when we didn't have the appropriate fields. This was causing
prawn to try to render this which was breaking:

`<data JobApplicationPdf::Table rows=[["How will you support the
school's ethos and aims?", nil], ["Are you currently following a
religion or faith?", {true: "Yes", false: "No"}]]>`

This PR fixes this issue by not rendering religious information for
vacancies with religion_type: "no_religion"