Set new flags to disable different job types
DisableExpensiveJobs was a broad flag used to enable/disable completely
unrelated areas.
EG: The ATS integrations and the subscription alerts where enabled or
disabled as a unit using this flag.

As a consequence, some test environments that needed to enable some of
the usually disabled jobs where triggering lots of exceptions due to
attempts to connet with external integrations.

To be able to enable/disable specific areas of our service jobs per
environment, we have split the flags in:

- DisableEmailNotifications
- DisableExpensiveJobs
- DisableIntegrations

We guard each job needed to be guarded/disabled in test envs using the
appropriate flag. And we will be able to enable, for example, email
notifications in a particular environment, while keeping the external
integrations disabled.