Align disable flag approach for Google Indx updates
The approach we follow in all the jobs is to have the disable flag used
as a guard clause within the job code itself. So the jobs get always
enqueued and the execution code skips if the job is disabled in the
environment.

For the google index update/delete jobs, we were guarding with the flag
on the job calls, what made the code/testing more complex and not
algined with the flag usage across the service.

Refactored it to behave as the other jobs.