Disable preview components outside dev/test envs.
The ViewComponent framework configuration in our service was enabling
the Preview feature in all the environments.

We do not need component previews to be accessible outside dev/test
environments (the default behaviour). So removing the config override.

We had no component preview to list anyway, but this closes the route
from public access in non dev environments.