Refactor equal opportunities report publication (#7923)
## Trello card URL

## Changes in this PR:
[1922](https://trello.com/c/gSXp59bc)

After clarifying some point with Brendan, currently we are send the
equal opportunities report data via the standard dfe analytics
mechanism; create/update .. events on the entity.
This original concern behind the creation of this custom event, was to
suppress the ability to match submitted job application with equal
opportunities report in the analytics. It is currently possible to match
the 2 entities based on request_uuid.

* Deleted stream_equal_opportunities_report_publication_job
* Removed custom event equal_opportunities_report_published
* Use EqualOpportunitiesReportUpdateJob to trigger data publication with
analytic event reqest_uuid different from the application submission