Ensure personal statement section is not marked as complete when impo… (#7948)
Prior to this change when we imported a personal statement from a
previous job application the personal statement was tagged as completed
so jobseekers did not need to review it. This change ensures it is not
tagged as complete so jobseekers have to review their personal statement
for every new application.