Allow publishers to opt out following a link
Publishers, signed in or not, will be able to set their "email_opt_out"
flag to "enabled" when following a particular URL and confirming their
wish.

We are allowing non-signed-in publishers to set this too for their
account as Publishers may have been removed from their organisation in
DfE Sign-in and lost access to our service. But they would be still
listed as publishers in our DB as they could be associated with
Vacancies in the service.

This will warrant if they receive a one-off communication email with a
link to the "Opt Out", they could still use it.

We use signed ids to avoid random access to this path by other users.

We have created a helper to generate "Opt Out" links in future email
communications.