Refactor publisher vacancies job application tag (#7942)
## Trello card URL
[1866](https://trello.com/c/ViQIH92Z)

## Changes in this PR:

This is extracted from #7876 as a basis for changes coming to that part
of the service.
No functional change is added at this stage.
Second PR #7943 contains the functional changes

* Added test coverage around job application status rendering both
publisher and jobseeker
* Added test coverage around job application status transitions
* Fix little bug when no status is select on the tag page

Some spec are pending and commented out for now as the PR #7918 is yet
to be merged

---------

Co-authored-by: Kyle MacPherson <kaimac2010@gmail.com>
Co-authored-by: Marc Sardón <marc.sardon@education.gov.uk>