Fix race condition in Vacancy analytics source updates
When a vacancy visit is tracked in Redis, and later the Redis updates
written in DB, the vacancy associated with the tracked visit may have
been removed from our system in between.

So far this has only happened in Review Apps (where the chances of
destroying Vacancies while testing and setting vacancies are higher),
but it could still happen in production.

Resolving it by ensuring the vacancy we're trying to register analytics
about still exists before trying to record the analytics in DB.