Disable MyNewTerm legacy integration
MyNewTerm is moving to the new Publisher ATS API integration.

We need to stop this integration from running automatically so there are
no clashes in future vacancies posted by them, but we may need to still
re-enable this.

So doing the minimum change to disable their legacy automated integration
until we confirm is safe to completely remove the rest of the code.