Little sanity refactor on PublisherPreference creation
This is not meant to fully resolve how the PublisherPreferences are
designed and implemented.

It just refactors the code in the publisher vacancies controller that is
currenly retrieving/setting the publisher preference.

The code was hard to read as the related logic was split.
Grouping it into a single method/block of code with explicit branching
conditions, I aim to help devs to be able to quickly read and understand
what is happening in the controller, and allow further improvements on
the preferences setup.