Little sanity refactor on PublisherPreference creation (#7993)
## Trello card URL

## Changes in this PR:

This is not meant to fully resolve how the PublisherPreferences are
designed and implemented.

It just refactors the code in the publisher vacancies controller that is
currently retrieving/setting the publisher preference.

The code was hard to read as the related logic was split. Grouping it
into a single method/block of code with explicit branching conditions, I
aim to help devs to be able to quickly read and understand what is
happening in the controller, and allow further improvements on the
preferences setup.

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>