Message archiving (#8020)
- https://trello.com/c/EaN3qvC8/2040-comms-build-ats-message-dashboard
-
https://trello.com/c/MDSWL2pL/2041-comms-build-ats-archive-functionality-part-of-message-dashboard

This PR introduces the MVP of the messages dashboard and allows
publishers to archive messages. This does not include any searching or
filtering features.

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>