Fix document virus check tempfile download path
The document virus check process was using the project root as
destination for the scanned file to be downloaded prior to be checked
for safety and deleted.

This was inocuous until our infra team tried to change the permissions
for our docker images and the file downloads started failing.

Resolving it by downloading the files where they belong: The project
"tmp" folder.
This folder has explicit permissions with the new docker user
permissions approach.