Fix document virus check tempfile download path (#8073)
## Trello card URL
- Related to: 
  - https://github.com/DFE-Digital/teaching-vacancies/pull/8069
  - https://github.com/DFE-Digital/teaching-vacancies/pull/8059
  -  
## Changes in this PR:

The document virus check process was using the project root as
destination for the scanned file to be downloaded prior to be checked
for safety and deleted.

This was innocuous until our infra team tried to change the permissions
for our Docker images, and the file uploads started failing during the
antivirus check.

Resolving it by downloading the files where they belong: The project
"tmp" folder.
This folder has explicit permissions with the new Docker user
permissions approach.

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>