Record Subscriptions Alerts stats into Sentry (#8114)
## Trello card URL

https://trello.com/c/CQjs31iq/2219-subscription-alert-runs-log-each-runtime-to-sentry

## Changes in this PR:

For each run of the Daily and Weekly alert email jobs, the job will send 
an info event to Sentry containing the runtime of the job, how many
emails were sent and how many vacancies the subscribers were alerted
about.

The new contextual info will be added to the `Context` Sentry block.