Avoid reloading subscriptions from DB 1 by 1
The guarding for the subscription email being removed while the job is
running results into over 100k individual DB reads.

This is caused by the "subscription.reload" call in the loop.

We can avoid this by moving that "fresh last second from DB" check
responsibility to the mailer itsef. As the mailer runs asynchronously
and is already loading the subscription from DB to send the email to the
subscriber.