Only do subscription location filtering if needed
The location handling was happening regardless of if any vacancy was
left after the non-location criteria has been applied.

This caused a lot of expensive computation calculations happening for
tens of thousands of subscriptions. When they should be just skipped,
since the subscriptions would have not match any vacancy by their search
criteria, independently on the location of the vacancies.

Now we only run the location filtering/computations if the subscription
has vacancy matches on the subscription criteria, and needs to calculate
if the location criteria also matches.