Add task to normalizing phases in subscriptions
Go through all the subscriptions in our DB and replace legacy values for
the current phases values.

This aims to reduce the computation and memory needed to do the mappings
over every single subscription on each job alerts run.

If it is already mapped at the source, we can avoid it on runtime and
just focus on matching.