Precompute filter lambdas
To reduce the number of memory allocations, we extract as much as
possible outside the vacancies loop for each subscription:

Precomputing the filter lambdas int an array once outside the vacancies
loop. Instead of looking up and calling each filter for each criterion
and vacancy.

This also reduces the hash lookups and block allocations.