Add Job to backfill subscription location fields
This is a very heavy backfill, so adding it as a job that can be rune
asynchronously in a worker.

The job:
1. obtains all the unique location & radius combinations from the search
criteria contained in the subscriptions table.
2. for each unique combination, tries to find a matching location
polygon.
3. if the polygon is found, buffers the polygon area by the given radius
and stores it as geometry data for all the subscriptions with that
location&radius.
4. if no polygon is found, retrieves the geopoint for that location and
stores it as geometry data for all the subscriptions with that location
& radius.
5. if no match is found by polygon or geopoint, Do not upgrade the
subscriptions.