Async set location data on subscription creation
The computation and set of location data for the subscription for the
given search criteria location requires either Polygon area buffering
computations or coordinates retrieval from our cache or Google API.

Wanted to extract this from the request call and execute it
asynchronously, so the user requests don't get impacted by it.

Originally tried to trigger it from a model callback, but avoiding it
running by default on test environment turned quite complex to achueve.

Decided to move towards a service object, extracting bits from the
controller logic related to the Subscription creation.