Fix buffered area computation for subscriptions
The issue

The previous buffered_geometry_area call was using SRID 3857
(Web Mercator projection) for calculating the buffering distance in meters.

While trying to use the result of this buffer testing the subscriptions
vacancies matching location filters, I noticed that a subscription using
"Basildon" polygon with a buffer of 50miles wasn't matching a vacancy
located in St. Albans (37 miles away from Basildon).

Then I started reading more about the SRID 3857 and found out that it distorts
distances and areas when moving away from the equator (the further the worse),
causing the distances needing a multiplier between 1.5x to 1.7x (needs more the
further north we go) to be semi-accurate.

The solution

Due to performance issues, we don't want to have to use geographic calculations
(in meters and very accurate) for buffering and distances, but we can use a
projected coordinate system in meters that is way more accurate for UK
locations: The British National Grid SRID (EPSG:27700).

It allows us to still use Geometry operations, be almost as accurate as
Geography and way faster.

Added a job to also backfill the subscriptions containing an area, replacing
their existing area with the recomputed area.